Update README for Bank Capital project
The README.md file has been updated to better represent the Bank Capital project. It now includes a project description, setup instructions, usage details, a list of technologies used and other relevant information. This will guide users and developers interested in the project on how to properly set up and use the application.